splitting *_sys.c files
