+++ source/null/clipboard_null.c
@@ -1,0 +1,24 @@
+#include "../client/client.h"
+
+/*
+* Sys_GetClipboardData
+*/
+char *Sys_GetClipboardData( qboolean primary )
+{
+	return NULL;
+}
+
+/*
+* Sys_SetClipboardData
+*/
+qboolean Sys_SetClipboardData( char *data )
+{
+	return qfalse;
+}
+
+/*
+* Sys_FreeClipboardData
+*/
+void Sys_FreeClipboardData( char *data )
+{
+}
