+++ source/qfusion_server.vcxproj
@@ -205,6 +205,7 @@
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="null\clipboard_null.c" />
     <ClCompile Include="qcommon\anticheat.c" />
     <ClCompile Include="qcommon\ascript.c" />
     <ClCompile Include="qalgo\base64.c" />
@@ -252,6 +253,7 @@
     <ClCompile Include="qcommon\svnrev.c" />
     <ClCompile Include="qalgo\q_trie.c" />
     <ClCompile Include="qcommon\webdownload.c" />
+    <ClCompile Include="win32\win_console.c" />
     <ClCompile Include="win32\win_fs.c" />
     <ClCompile Include="win32\win_net.c" />
     <ClCompile Include="win32\win_sys.c" />
@@ -260,6 +262,7 @@
     <ClCompile Include="win32\win_lib.c" />
     <ClCompile Include="win32\win_threads.c" />
     <ClCompile Include="null\sys_vfs_null.c" />
+    <ClCompile Include="win32\win_time.c" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="angelwrap\qas_public.h" />
