+++ source/qfusion_server.vcxproj.filters
@@ -180,6 +180,15 @@
     <ClCompile Include="qcommon\steam.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="win32\win_time.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="win32\win_console.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="null\clipboard_null.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="qcommon\anticheat.h">
