+++ source/server/CMakeLists.txt
@@ -36,6 +36,7 @@ file(GLOB SERVER_SOURCES
     "../qcommon/steam.c"
     "*.c"
     "../null/cl_null.c"
+    "../null/clipboard_null.c"
     "../gameshared/q_*.c"
     "../qalgo/*.c"
     "../matchmaker/mm_*.c"
@@ -46,6 +47,8 @@ if (${CMAKE_SYSTEM_NAME} MATCHES "Windows")
         "../win32/win_fs.c"
         "../win32/win_net.c"
         "../win32/win_sys.c"
+        "../win32/win_console.c"
+        "../win32/win_time.c"
         "../win32/win_lib.c"
         "../win32/win_threads.c"
         "../null/sys_vfs_null.c"
@@ -59,6 +62,8 @@ else()
         "../unix/unix_fs.c"
         "../unix/unix_net.c"
         "../unix/unix_sys.c"
+        "../unix/unix_console.c"
+        "../unix/unix_time.c"
         "../unix/unix_lib.c"
         "../unix/unix_threads.c"
         "../null/sys_vfs_null.c"
