+++ source/tv_server/CMakeLists.txt
@@ -34,6 +34,7 @@ file(GLOB TV_SERVER_SOURCES
     "../null/cl_null.c"
     "../null/ascript_null.c"
     "../null/mm_null.c"
+    "../null/clipboard_null.c"
     "../gameshared/q_*.c"
     "../qalgo/*.c"
     "tv_module/*.c"
@@ -44,6 +45,8 @@ if (${CMAKE_SYSTEM_NAME} MATCHES "Windows")
         "../win32/win_fs.c"
         "../win32/win_net.c"
         "../win32/win_sys.c"
+        "../win32/win_console.c"
+        "../win32/win_time.c"
         "../win32/win_lib.c"
         "../win32/win_threads.c"
         "../null/sys_vfs_null.c"
@@ -57,6 +60,8 @@ else()
         "../unix/unix_fs.c"
         "../unix/unix_net.c"
         "../unix/unix_sys.c"
+        "../unix/unix_console.c"
+        "../unix/unix_time.c"
         "../unix/unix_lib.c"
         "../unix/unix_threads.c"
         "../null/sys_vfs_null.c"
