+++ source/unix/unix_sys.c
@@ -53,12 +53,11 @@ FIXME:  This will be remidied once a native Mac port is complete
 #include "glob.h"
 
 cvar_t *nostdout;
-static qboolean nostdout_backup_val = qfalse;
+qboolean nostdout_backup_val = qfalse;
 
 unsigned sys_frame_time;
 
 uid_t saved_euid;
-qboolean stdin_active = qtrue;
 
 // =======================================================================
 // General routines
@@ -115,82 +114,6 @@ static void InitSig( void )
 	signal( SIGSEGV, signal_handler );
 	signal( SIGTERM, signal_handler );
 	signal( SIGINT, signal_handler );
-}
-
-static void Sys_AnsiColorPrint( const char *msg )
-{
-	static char buffer[2096];
-	int         length = 0;
-	static int  q3ToAnsi[ 8 ] =
-	{
-		30, // COLOR_BLACK
-		31, // COLOR_RED
-		32, // COLOR_GREEN
-		33, // COLOR_YELLOW
-		34, // COLOR_BLUE
-		36, // COLOR_CYAN
-		35, // COLOR_MAGENTA
-		0   // COLOR_WHITE
-	};
-
-	while( *msg )
-	{
-		char c = *msg;
-		int colorindex;
-
-		int gc = Q_GrabCharFromColorString( &msg, &c, &colorindex );
-		if( gc == GRABCHAR_COLOR || (gc == GRABCHAR_CHAR && c == '\n') )
-		{
-			// First empty the buffer
-			if( length > 0 )
-			{
-				buffer[length] = '\0';
-				fputs( buffer, stdout );
-				length = 0;
-			}
-
-			if( c == '\n' )
-			{
-				// Issue a reset and then the newline
-				fputs( "\033[0m\n", stdout );
-			}
-			else
-			{
-				// Print the color code
-				Q_snprintfz( buffer, sizeof( buffer ), "\033[%dm", q3ToAnsi[ colorindex ] );
-				fputs( buffer, stdout );
-			}
-		}
-		else if( gc == GRABCHAR_END )
-			break;
-		else
-		{
-			if( length >= sizeof( buffer ) - 1 )
-				break;
-			buffer[length++] = c;
-		}
-	}
-
-	// Empty anything still left in the buffer
-	if( length > 0 )
-	{
-		buffer[length] = '\0';
-		fputs( buffer, stdout );
-	}
-}
-
-void Sys_ConsoleOutput( char *string )
-{
-	if( nostdout && nostdout->integer )
-		return;
-	if( nostdout_backup_val )
-		return;
-
-#if 0
-	fputs( string, stdout );
-#else
-	Sys_AnsiColorPrint( string );
-#endif
 }
 
 /*
@@ -257,98 +180,11 @@ void Sys_Error( const char *format, ... )
 }
 
 /*
-* Sys_Microseconds
-*/
-static unsigned long sys_secbase;
-quint64 Sys_Microseconds( void )
-{
-	struct timeval tp;
-	struct timezone tzp;
-
-	gettimeofday( &tp, &tzp );
-
-	if( !sys_secbase )
-	{
-		sys_secbase = tp.tv_sec;
-		return tp.tv_usec;
-	}
-
-	// TODO handle the wrap
-	return (quint64)( tp.tv_sec - sys_secbase )*1000000 + tp.tv_usec;
-}
-
-/*
-* Sys_Milliseconds
-*/
-unsigned int Sys_Milliseconds( void )
-{
-	return Sys_Microseconds() / 1000;
-}
-
-/*
-* Sys_XTimeToSysTime
-* 
-* Sub-frame timing of events returned by X
-* Ported from Quake III Arena source code.
-*/
-int Sys_XTimeToSysTime( unsigned long xtime )
-{
-	int ret, time, test;
-
-	// some X servers (like suse 8.1's) report weird event times
-	// if the game is loading, resolving DNS, etc. we are also getting old events
-	// so we only deal with subframe corrections that look 'normal'
-	ret = xtime - (unsigned long)(sys_secbase * 1000);
-	time = Sys_Milliseconds();
-	test = time - ret;
-
-	if( test < 0 || test > 30 ) // in normal conditions I've never seen this go above
-		return time;
-	return ret;
-}
-
-/*
 * Sys_Sleep
 */
 void Sys_Sleep( unsigned int millis )
 {
 	usleep( millis * 1000 );
-}
-
-char *Sys_ConsoleInput( void )
-{
-	static char text[256];
-	int len;
-	fd_set fdset;
-	struct timeval timeout;
-
-	if( !dedicated || !dedicated->integer )
-		return NULL;
-
-	if( !stdin_active )
-		return NULL;
-
-	FD_ZERO( &fdset );
-	FD_SET( 0, &fdset ); // stdin
-	timeout.tv_sec = 0;
-	timeout.tv_usec = 0;
-	if( select( 1, &fdset, NULL, NULL, &timeout ) == -1 || !FD_ISSET( 0, &fdset ) )
-		return NULL;
-
-	len = read( 0, text, sizeof( text ) );
-	if( len == 0 )
-	{           // eof!
-		Com_Printf( "EOF from stdin, console input disabled...\n" );
-		stdin_active = qfalse;
-		return NULL;
-	}
-
-	if( len < 1 )
-		return NULL;
-
-	text[len-1] = 0; // rip off the /n and terminate
-
-	return text;
 }
 
 /*
@@ -387,33 +223,6 @@ void Sys_SendKeyEvents( void )
 	// grab frame time
 	sys_frame_time = Sys_Milliseconds();
 }
-
-#ifdef DEDICATED_ONLY
-
-/*
-* Sys_GetClipboardData
-*/
-char *Sys_GetClipboardData( qboolean primary )
-{
-	return NULL;
-}
-
-/*
-* Sys_SetClipboardData
-*/
-qboolean Sys_SetClipboardData( char *data )
-{
-	return qfalse;
-}
-
-/*
-* Sys_FreeClipboardData
-*/
-void Sys_FreeClipboardData( char *data )
-{
-}
-
-#endif
 
 #ifndef __APPLE__
 /*
