+++ source/win32/win_sys.c
@@ -40,7 +40,7 @@ int ActiveApp;
 int Minimized;
 int AppFocused;
 
-static HANDLE hinput, houtput;
+HANDLE hinput, houtput;
 
 unsigned sys_msg_time;
 unsigned sys_frame_time;
@@ -53,10 +53,11 @@ char *argv[MAX_NUM_ARGVS];
 static dynvar_get_status_t Sys_GetAffinity_f( void **affinity );
 static dynvar_set_status_t Sys_SetAffinity_f( void *affinity );
 
-static qboolean	hwtimer;
-static dynvar_t	*hwtimer_var;
-static int milli_offset = 0;
-static qint64 micro_offset = 0;
+extern qboolean	hwtimer;
+extern dynvar_t	*hwtimer_var;
+extern qint64 hwtimer_freq;
+extern int milli_offset;
+extern qint64 micro_offset;
 
 static dynvar_get_status_t Sys_GetHwTimer_f( void **val );
 static dynvar_set_status_t Sys_SetHwTimer_f( void *val );
@@ -110,82 +111,6 @@ void Sys_Quit( void )
 
 //================================================================
 
-/*
-* Sys_Milliseconds
-*/
-// wsw: pb adapted High Res Performance Counter code from ezquake
-static qint64 freq;
-
-static void Sys_InitTime( void )
-{
-	char *hwtimerStr;
-	assert( hwtimer_var );
-	Dynvar_GetValue( hwtimer_var, (void **)&hwtimerStr );
-	assert( hwtimerStr );
-	if( hwtimerStr[0] == '0' && COM_CheckParm( "-hwtimer" ) )
-	{
-		// hwtimer set by command line parameter (deprecated)
-		Dynvar_SetValue( hwtimer_var, "1" );
-	}
-	// finally check whether hwtimer is activated
-	if( hwtimer )
-		Com_Printf( "Using High Resolution Performance Counter\n" );
-	else
-		Com_Printf( "Using timeGetTime\n" );
-}
-
-static unsigned int Sys_Milliseconds_TGT( void )
-{
-	static unsigned int base;
-	static qboolean	initialized = qfalse;
-	unsigned int now;
-
-	if( !initialized )
-	{
-		// let base retain 16 bits of effectively random data which is used
-		//for quickly generating random numbers
-		base = timeGetTime() & 0xffff0000;
-		initialized = qtrue;
-	}
-
-	now = timeGetTime();
-
-	return now - base;
-}
-
-static quint64 Sys_Microseconds_QPC( void )
-{
-	static qboolean first = qtrue;
-	static qint64 p_start;
-
-	qint64 p_now;
-	QueryPerformanceCounter( (LARGE_INTEGER *) &p_now );
-
-	if( first )
-	{
-		first = qfalse;
-		p_start = p_now;
-	}
-
-	return ( ( p_now - p_start ) * 1000000 ) / freq;
-}
-
-unsigned int Sys_Milliseconds( void )
-{
-	if( hwtimer )
-		return ( Sys_Microseconds_QPC() + micro_offset ) / 1000;
-	else
-		return Sys_Milliseconds_TGT() + milli_offset;
-}
-
-quint64 Sys_Microseconds( void )
-{
-	if( hwtimer )
-		return Sys_Microseconds_QPC() + micro_offset;
-	else
-		return (quint64)( Sys_Milliseconds_TGT() + milli_offset ) * 1000;
-}
-
 void Sys_Sleep( unsigned int millis )
 {
 	Sleep( millis );
@@ -237,7 +162,7 @@ void Sys_InitDynvars( void )
 	char *dummyStr;
 	dynvar_t *affinity_var;
 
-	QueryPerformanceFrequency( (LARGE_INTEGER *) &freq );
+	QueryPerformanceFrequency( (LARGE_INTEGER *) &hwtimer_freq );
 
 	affinity_var = Dynvar_Create( "sys_affinity", qtrue, Sys_GetAffinity_f, Sys_SetAffinity_f );
 	assert( affinity_var );
@@ -250,185 +175,6 @@ void Sys_InitDynvars( void )
 	Dynvar_AddListener( hwtimer_var, Sys_SynchronizeTimers_f );
 	Dynvar_SetValue( hwtimer_var, "0" );
 }
-
-#define MAX_CONSOLETEXT 256
-static char console_text[MAX_CONSOLETEXT];
-static int console_textlen;
-
-static char *OEM_to_utf8( const char *str )
-{
-	WCHAR wstr[MAX_CONSOLETEXT];
-	static char utf8str[MAX_CONSOLETEXT*4]; /* longest valid utf8 sequence is 4 bytes */
-
-	MultiByteToWideChar( CP_OEMCP, 0, str, -1, wstr, sizeof( wstr )/sizeof( WCHAR ) );
-	wstr[sizeof( wstr )/sizeof( wstr[0] ) - 1] = 0;
-	WideCharToMultiByte( CP_UTF8, 0, wstr, -1, utf8str, sizeof( utf8str ), NULL, NULL );
-	utf8str[sizeof( utf8str ) - 1] = 0;
-
-	return utf8str;
-}
-
-static char *utf8_to_OEM( const char *utf8str )
-{
-	WCHAR wstr[MAX_PRINTMSG];
-	static char oemstr[MAX_PRINTMSG];
-
-	MultiByteToWideChar( CP_UTF8, 0, utf8str, -1, wstr, sizeof( wstr )/sizeof( WCHAR ) );
-	wstr[sizeof( wstr )/sizeof( wstr[0] ) - 1] = 0;
-	WideCharToMultiByte( CP_OEMCP, 0, wstr, -1, oemstr, sizeof( oemstr ), "?", NULL );
-	oemstr[sizeof( oemstr ) - 1] = 0;
-
-	return oemstr;
-}
-
-/*
-* Sys_ConsoleInput
-*/
-char *Sys_ConsoleInput( void )
-{
-	INPUT_RECORD rec;
-	int ch;
-	DWORD dummy;
-	DWORD numread, numevents;
-
-	if( !dedicated || !dedicated->integer )
-		return NULL;
-
-	for(;; )
-	{
-		if( !GetNumberOfConsoleInputEvents( hinput, &numevents ) )
-			Sys_Error( "Error getting # of console events" );
-
-		if( numevents <= 0 )
-			break;
-
-		if( !ReadConsoleInput( hinput, &rec, 1, &numread ) )
-			Sys_Error( "Error reading console input" );
-
-		if( numread != 1 )
-			Sys_Error( "Couldn't read console input" );
-
-		if( rec.EventType == KEY_EVENT )
-		{
-			if( !rec.Event.KeyEvent.bKeyDown )
-			{
-				ch = rec.Event.KeyEvent.uChar.AsciiChar;
-
-				switch( ch )
-				{
-				case '\r':
-					WriteFile( houtput, "\r\n", 2, &dummy, NULL );
-
-					if( console_textlen )
-					{
-						console_text[console_textlen] = 0;
-						console_textlen = 0;
-						return OEM_to_utf8( console_text );
-					}
-					break;
-
-				case '\b':
-					if( console_textlen )
-					{
-						console_textlen--;
-						WriteFile( houtput, "\b \b", 3, &dummy, NULL );
-					}
-					break;
-
-				default:
-					if( ( unsigned char )ch >= ' ' )
-					{
-						if( console_textlen < sizeof( console_text )-2 )
-						{
-							WriteFile( houtput, &ch, 1, &dummy, NULL );
-							console_text[console_textlen] = ch;
-							console_textlen++;
-						}
-					}
-					break;
-				}
-			}
-		}
-	}
-
-	return NULL;
-}
-
-static void PrintColoredText( const char *s )
-{
-	char c;
-	int colorindex;
-	DWORD dummy;
-
-	while( *s )
-	{
-		int gc = Q_GrabCharFromColorString( &s, &c, &colorindex );
-		if( gc == GRABCHAR_CHAR )
-		{
-			if( c == '\n' )
-				SetConsoleTextAttribute( houtput, 7 );
-			// I hope it's not too slow to output char by char
-			WriteFile( houtput, &c, 1, &dummy, NULL );
-		}
-		else if( gc == GRABCHAR_COLOR )
-		{
-			switch( colorindex )
-			{
-			case 0: colorindex = 3; break;	// dark cyan instead of black to keep it visible
-			case 1: colorindex = 12; break;
-			case 2: colorindex = 10; break;
-			case 3: colorindex = 14; break;
-			case 4: colorindex = 9; break;
-			case 5: colorindex = 11; break;	// note that cyan and magenta are 
-			case 6: colorindex = 13; break;	// not where one might expect
-			case 8: colorindex = 6; break;
-			case 9: colorindex = 8; break;
-			default:
-			case 7: colorindex = 7; break;	// 15 would be bright white
-			};
-			SetConsoleTextAttribute( houtput, colorindex );
-		}
-		else if( gc == GRABCHAR_END )
-			break;
-		else
-			assert( 0 );
-	}
-}
-
-/*
-* Sys_ConsoleOutput
-* 
-* Print text to the dedicated console
-*/
-void Sys_ConsoleOutput( char *string )
-{
-	DWORD dummy;
-	char text[MAX_CONSOLETEXT+2];	/* need 2 chars for the \r's */
-
-	if( !dedicated || !dedicated->integer )
-		return;
-
-	if( console_textlen )
-	{
-		text[0] = '\r';
-		memset( &text[1], ' ', console_textlen );
-		text[console_textlen+1] = '\r';
-		text[console_textlen+2] = 0;
-		WriteFile( houtput, text, console_textlen+2, &dummy, NULL );
-	}
-
-	string = utf8_to_OEM( string );
-
-#if 0
-	WriteFile( houtput, string, (unsigned)strlen( string ), &dummy, NULL );
-#else
-	PrintColoredText( string );
-#endif
-
-	if( console_textlen )
-		WriteFile( houtput, console_text, console_textlen, &dummy, NULL );
-}
-
 
 /*
 * myTranslateMessage
@@ -472,89 +218,6 @@ void Sys_SendKeyEvents( void )
 
 	// grab frame time
 	sys_frame_time = timeGetTime(); // FIXME: should this be at start?
-}
-
-
-
-/*
-* Sys_GetClipboardData
-*/
-char *Sys_GetClipboardData( qboolean primary )
-{
-	char *utf8text = NULL;
-	int utf8size;
-	WCHAR *cliptext;
-
-	if( OpenClipboard( NULL ) != 0 )
-	{
-		HANDLE hClipboardData;
-
-		if( ( hClipboardData = GetClipboardData( CF_UNICODETEXT ) ) != 0 )
-		{
-			if( ( cliptext = GlobalLock( hClipboardData ) ) != 0 )
-			{
-				utf8size = WideCharToMultiByte( CP_UTF8, 0, cliptext, -1, NULL, 0, NULL, NULL );
-				utf8text = Q_malloc( utf8size );
-				WideCharToMultiByte( CP_UTF8, 0, cliptext, -1, utf8text, utf8size, NULL, NULL );
-				GlobalUnlock( hClipboardData );
-			}
-		}
-		CloseClipboard();
-	}
-	return utf8text;
-}
-
-/*
-* Sys_SetClipboardData
-*/
-qboolean Sys_SetClipboardData( char *data )
-{
-	size_t size;
-	HGLOBAL hglbCopy;
-	UINT uFormat;
-	LPCSTR cliptext = data;
-	LPWSTR lptstrCopy;
-
-	// open the clipboard, and empty it
-	if( !OpenClipboard( NULL ) ) 
-		return qfalse;
-
-	EmptyClipboard();
-
-	size = MultiByteToWideChar( CP_UTF8, 0, cliptext, -1, NULL, 0 );
-
-	// allocate a global memory object for the text
-	hglbCopy = GlobalAlloc( GMEM_MOVEABLE, (size + 1) * sizeof( *lptstrCopy ) ); 
-	if( hglbCopy == NULL )
-	{
-		CloseClipboard(); 
-		return qfalse; 
-	} 
-
-	// lock the handle and copy the text to the buffer
-	lptstrCopy = GlobalLock( hglbCopy ); 
-
-	uFormat = CF_UNICODETEXT;
-	MultiByteToWideChar( CP_UTF8, 0, cliptext, -1, lptstrCopy, size );
-	lptstrCopy[size] = 0;
-
-	GlobalUnlock( hglbCopy ); 
-
-	// place the handle on the clipboard
-	SetClipboardData( uFormat, hglbCopy );
-
-	// close the clipboard
-	CloseClipboard();
-
-	return qtrue;
-}
-
-/*
-* Sys_FreeClipboardData
-*/
-void Sys_FreeClipboardData( char *data )
-{
-	Q_free( data );
 }
 
 /*
@@ -815,7 +478,7 @@ static dynvar_set_status_t Sys_SetHwTimer_f( void *val )
 		hwtimer = 0;
 		return DYNVAR_SET_OK;
 	case '1':
-		if( freq )
+		if( hwtimer_freq )
 		{
 			hwtimer = 1;
 			return DYNVAR_SET_OK;
